requirements.txt
streamlit
ultralytics
mediapipe
wikipedia-api
gtts
opencv-python
pillow
numpy
torch
torchvision
torchaudio
opencv-python-headless
ultralytics
